requirements.txt
fastapi==0.115.12
sqlmodel==0.0.24
psycopg2==2.9.10
python-dotenv==1.1.0
uvicorn==0.34.0
pytest==8.3.5
httpx==0.28.1
ruff==0.11.2